JBPM-5420: Stunner - move BPMN specific property types into bpmn-api module
