+++ kie-wb-common-stunner/kie-wb-common-stunner-sets/kie-wb-common-stunner-bpmn/kie-wb-common-stunner-bpmn-api/src/main/java/org/kie/workbench/common/stunner/bpmn/definition/property/type/AssignmentsType.java
@@ -13,15 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.workbench.common.stunner.core.definition.property.type;
+package org.kie.workbench.common.stunner.bpmn.definition.property.type;
 
 import org.jboss.errai.common.client.api.annotations.Portable;
 import org.kie.workbench.common.stunner.core.definition.property.PropertyType;
 
 @Portable
-public class AssigneeType implements PropertyType {
+public class AssignmentsType implements PropertyType {
 
-    public static final String name = "stunner.assignee";
+    public static final String name = "stunner.assignments";
 
     @Override
     public String getName() {
@@ -33,10 +33,10 @@ public class AssigneeType implements PropertyType {
         if ( this == o ) {
             return true;
         }
-        if ( !( o instanceof AssigneeType ) ) {
+        if ( !( o instanceof AssignmentsType ) ) {
             return false;
         }
-        AssigneeType that = ( AssigneeType ) o;
+        AssignmentsType that = ( AssignmentsType ) o;
         if ( !name.equals( that.name ) ) {
             return false;
         }
@@ -50,7 +50,7 @@ public class AssigneeType implements PropertyType {
 
     @Override
     public String toString() {
-        return "AssigneeType{" +
+        return "AssignmentsType{" +
                 "name='" + name + '\'' +
                 '}';
     }
