+++ kie-wb-common-stunner/kie-wb-common-stunner-sets/kie-wb-common-stunner-bpmn/kie-wb-common-stunner-bpmn-backend/src/main/java/org/kie/workbench/common/stunner/bpmn/backend/marshall/json/oryx/property/VariablesTypeSerializer.java
@@ -17,7 +17,7 @@
 package org.kie.workbench.common.stunner.bpmn.backend.marshall.json.oryx.property;
 
 import org.kie.workbench.common.stunner.core.definition.property.PropertyType;
-import org.kie.workbench.common.stunner.core.definition.property.type.VariablesType;
+import org.kie.workbench.common.stunner.bpmn.definition.property.type.VariablesType;
 
 import javax.enterprise.context.ApplicationScoped;
